restructuring directories
