+++ Frontend/package.json
@@ -1,10 +1,10 @@
 {
   "scripts": {
     "dev": "next dev",
-    "lint": "npx eslint ./pages && echo Linting was successful. No errors were found. || echo Linting failed. Errors were found.",
+    "lint": "npx eslint ./src && echo Linting was successful. No errors were found. || echo Linting failed. Errors were found.",
     "test": "echo 'No tests specified to run'",
-    "pretty": "npx prettier --write \"./pages/**.{ts,tsx}\"",
-    "pretty-check": "npx prettier --check \"./pages/**.{ts,tsx}\""
+    "pretty": "npx prettier --write \"./src/**/*.{ts,tsx}\"",
+    "pretty-check": "npx prettier --check \"./src/**.{ts,tsx}\""
   },
   "dependencies": {
     "bootstrap": "^5.3.3",
