+++ Frontend/src/config/settings.ts
@@ -1,0 +1,3 @@
+export class Settings {
+    static readonly SERVER_URL = "http://localhost:8080";
+}
