+++ Frontend/src/types/types.ts
@@ -1,0 +1,6 @@
+export enum PageType {
+    QUIZ_TYPE_SELECTION = "quiz_type_selection",
+    QUIZ_LEVEL_SELECTION = "quiz_level_selection",
+    QUIZ = "quiz",
+    RESULTS = "results",
+}
